Update 1
Okay, so this is the first update. Just made a basic update to how the GUI looks, made the button bigger (I know it looks like shit I'll change it in the tomorrows update (If I'll be free tomorrow)) and made the dropdown show the quality instead of the resolutions. It'll look more organized this way.